Concept Push 1.0
Ok so this was a weird one. Had this idea while I was extremely sick and was bored out of my mind at home. Looking back, not quite sure what I was trying to achieve, but the idea was custom triggers based on arm movement. Final goal was to have the app connect to some sort of web based program that had a ball. As the user moved their arm up the ball would rise emulating the "Force" in Star Wars.

If you raise your arm up, the app triggers something. That was as far as I got with this project. Will most likely never get finished, not as important to non-sick Taylor.